Don't send notifications when the user is driving
