+++ ios/App/LocationManager.swift
@@ -88,6 +88,11 @@ class LocationManager: NSObject, CLLocationManagerDelegate {
 
   func locationManager(_ manager: CLLocationManager, didEnterRegion region: CLRegion) {
     if region is CLCircularRegion {
+      if MotionManager.shared.currentActivity?.automotive == true {
+        print("Currently driving")
+        return
+      }
+
       let identifier = region.identifier
       var identifiers = NotificationManager.shared.identifiers
       let sendAgainAt = identifiers[identifier]
@@ -100,7 +105,7 @@ class LocationManager: NSObject, CLLocationManagerDelegate {
 
           PlaceManager.contentForPlace(place: place) { (content) in
             let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: nil)
-            let center = UNUserNotificationCenter.current()            
+            let center = UNUserNotificationCenter.current()
             center.add(request)
         }
       }
